+++ frontend/src/components/input/CustomTextInput.jsx
@@ -1,4 +1,8 @@
+import { useAppContext } from '../../lib/context';
+
 export default function CustomTextInput({ value, onChange, placeholder, type, ...props }) {
+    const context = useAppContext();
+
     return (
         <div
             className='d-flex flex-column justify-content-start align-items center rounded-lg'
@@ -10,6 +14,7 @@ export default function CustomTextInput({ value, onChange, placeholder, type, ..
                 value={value}
                 onChange={e => onChange(e.target.value)}
                 placeholder={placeholder}
+                style={{backgroundColor: context.colorScheme.secondaryBackground, color: context.colorScheme.textColor}}
             />
         </div>
     );
