+++ frontend/src/components/input/CustomTextarea.jsx
@@ -1,6 +1,5 @@
 import { useAppContext } from '../../lib/context';
 
-
 export default function CustomTextarea({ value, onChange, ...props }) {
     const context = useAppContext();
