+++ frontend/src/components/profile/FriendsTab.jsx
@@ -1,19 +1,37 @@
+// this is the friends tab component
+// it displays the friends tab for the profile page
+
 import { Link } from 'react-router-dom';
 import { useState, useEffect } from 'react';
+
 import DefaultPFP from '../../assets/default-pfp.jpg';
+
 import Loading from '../utility/Loading';
 import Alert from '../utility/Alert';
+
 import api from '../../lib/api';
+import { useAppContext } from '../../lib/context';
 
 
 function UserResult({ user, handleRemoveFriend, showAddButton }) {
     const [error, setError] = useState(null);
+    const context = useAppContext();
 
     if (error) return <Alert message={error} hideAlert={() => setError(null)}/>
 
     return (
-        <div className='p-2 rounded-lg mb-2 bg-light d-flex flex-row justify-content-between align-items-center'>
-            <Link className='d-flex flex-row align-items-center' to={'/profile/' + user.username}>
+        <div 
+            className='p-2 rounded-lg mb-2 d-flex flex-row justify-content-between align-items-center'
+            style={{
+                backgroundColor: context.colorScheme.tertiaryBackground,
+                color: context.colorScheme.textColor
+            }}
+        >
+            <Link 
+                className='d-flex flex-row align-items-center' 
+                to={'/profile/' + user.username}
+                style={{ textDecoration: 'none' }}
+            >
                 <img
                     className='mr-2'
                     src={user.profile_picture || DefaultPFP}
@@ -21,19 +39,46 @@ function UserResult({ user, handleRemoveFriend, showAddButton }) {
                         width: '40px',
                         borderRadius: 1000,
                     }}
+                    alt={user.username}
                 />
                 <div className='d-flex flex-column'>
                     {(user.name && user.name !== '') ?
-                        <p className='m-0 font-weight-bold text-dark'>{user.name}</p>
+                        <p 
+                            className='m-0 font-weight-bold'
+                            style={{ color: context.colorScheme.textColor }}
+                        >
+                            {user.name}
+                        </p>
                     : null}
-                    <p className='m-0 text-muted text-dark'>@{user.username}</p>
+                    <p 
+                        className='m-0'
+                        style={{ color: context.colorScheme.secondaryText }}
+                    >
+                        @{user.username}
+                    </p>
                 </div>
             </Link>
             {showAddButton ?
                 <button
                     onClick={handleRemoveFriend}
-                    className='btn btn-danger'
-                >Remove</button>
+                    className='btn'
+                    style={{
+                        backgroundColor: context.colorScheme.danger,
+                        color: 'white',
+                        border: 'none',
+                        transition: 'background-color 0.2s ease, transform 0.1s ease',
+                    }}
+                    onMouseOver={(e) => {
+                        e.currentTarget.style.backgroundColor = `${context.colorScheme.danger}dd`;
+                        e.currentTarget.style.transform = 'scale(1.03)';
+                    }}
+                    onMouseOut={(e) => {
+                        e.currentTarget.style.backgroundColor = context.colorScheme.danger;
+                        e.currentTarget.style.transform = 'scale(1)';
+                    }}
+                >
+                    Remove
+                </button>
             : null}
         </div>
     );
@@ -44,6 +89,7 @@ export default function FriendsTab({ username }) {
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState(null);
     const [friends, setFriends] = useState(null);
+    const context = useAppContext();
 
     const getData = async () => {
         setLoading(true);
@@ -73,7 +119,7 @@ export default function FriendsTab({ username }) {
 
     if (friends) {
         return (
-            <div>
+            <div style={{ backgroundColor: context.colorScheme.secondaryBackground }}>
                 {friends.length > 0 ?
                     <div>
                         {friends.map((item, i) => (
@@ -87,7 +133,9 @@ export default function FriendsTab({ username }) {
                     </div>
                 :
                     <div className='p-5 d-flex flex-column justify-content-center align-items-center'>
-                        <p className='m-0'>Nothing to see here!</p>
+                        <p className='m-0' style={{ color: context.colorScheme.textColor }}>
+                            Nothing to see here!
+                        </p>
                     </div>
                 }
             </div>
