+++ frontend/src/components/utility/Loading.jsx
@@ -1,6 +1,4 @@
-
 import LoadingGIF from '../../assets/loading.gif';
-
 
 export default function Loading({ className, ...props }) {
     return (
