+++ frontend/src/components/utility/Modal.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
+import { useAppContext } from '../../lib/context';
 
 export default function Modal({ showModal, hideModal, children }) {
     if (!showModal) return null;
-    
+    const context = useAppContext();
     const handleModalContainerClick = (e) => {
         // Prevent clicks within the modal content from closing the modal
         e.stopPropagation();
@@ -22,13 +23,14 @@ export default function Modal({ showModal, hideModal, children }) {
             onClick={hideModal}
         >
             <div 
-                className="modal-content"
+                className="modal-content rounded shadow"
                 style={{
                     width: '600px',
                     maxWidth: '90%',
                     maxHeight: '90%',
                     overflow: 'auto',
                     zIndex: 1051,
+                    backgroundColor: context.colorScheme.secondaryBackground,
                 }}
                 onClick={handleModalContainerClick}
             >
