+++ frontend/src/components/utility/Sidebar.jsx
@@ -1,4 +1,5 @@
 import { useState, useEffect, useRef, useCallback } from 'react';
+
 import { useAppContext } from '../../lib/context';
