+++ frontend/src/pages/Event.jsx
@@ -1,6 +1,8 @@
 import { useParams } from 'react-router-dom';
+
 import EventHeader from '../components/events/EventHeader';
 import EventComments from '../components/events/EventComments';
+
 import { useAppContext } from '../lib/context';
 
 export default function Event() {
@@ -14,13 +16,14 @@ export default function Event() {
                 overflowY: 'auto', 
                 backgroundColor: context.colorScheme.backgroundColor, 
                 color: context.colorScheme.textColor,
-                height: 'calc(100vh - 48px)' // Adjust for navbar height
+                height: 'calc(100vh - 48px)' // for the navbar
             }}
         >
             <div
                 className='container my-3 w-100 rounded
                 d-flex flex-column justify-content-start align-items-stretch'
-                style={{backgroundColor: context.colorScheme.secondaryBackground}}
+                style={{backgroundColor: context.colorScheme.secondaryBackground,
+                    paddingTop: '15px'}}
             >
                 <div className='mb-3'>
                     <EventHeader eventId={id}/>
