+++ frontend/src/pages/Landing.jsx
@@ -1,0 +1,23 @@
+// this is the landing page component
+// it displays the landing page for the app
+
+import React from "react";
+import { Link } from "react-router-dom";
+
+import { useAppContext } from '../lib/context';
+
+export default function Landing() {
+    const context = useAppContext();
+    return (
+        <div style={{height: '100vh'}} className='d-flex justify-content-center align-items-center'>
+            <div className='d-flex flex-column align-items-center' style={{backgroundColor: context.colorScheme.backgroundColor, color: context.colorScheme.textColor}}>
+                <h1 className='h1'>CalendarMedia</h1>
+                <p>Look at what other people are doing!</p>
+                <div>
+                    <Link className='btn btn-primary mr-3' to='/login'>Login</Link>
+                    <Link className='btn btn-secondary' to='/register'>Register</Link>
+                </div>
+            </div>
+        </div>
+    );
+}
